that crazy ugly exec ai move logic now in mixins/gamepage/execaimove
